+++ package.json
@@ -30,6 +30,7 @@
     "@types/react-redux": "^7.1.25",
     "@types/react-slick": "^0.23.10",
     "@types/redux-logger": "^3.0.9",
+    "aws-sdk": "^2.1413.0",
     "axios": "^1.3.6",
     "date-fns": "^2.29.3",
     "dayjs": "^1.11.7",
@@ -39,8 +40,8 @@
     "husky": "^8.0.3",
     "lodash": "^4.17.21",
     "next": "13.2.4",
+    "quill-image-resize-module-ts": "^3.0.3",
     "react": "18.2.0",
-    "react-aws-s3-typescript": "^1.1.5",
     "react-datepicker": "^4.11.0",
     "react-dom": "18.2.0",
     "react-icons": "^4.8.0",
@@ -50,7 +51,8 @@
     "react-slick": "^0.29.0",
     "sass": "^1.61.0",
     "slick-carousel": "^1.8.1",
-    "typescript": "5.0.2"
+    "typescript": "5.0.2",
+    "uuid": "^9.0.0"
   },
   "devDependencies": {
     "@emotion/jest": "^11.10.5",
@@ -59,6 +61,7 @@
     "@testing-library/react": "^14.0.0",
     "@testing-library/react-hooks": "^8.0.1",
     "@testing-library/user-event": "^14.4.3",
+    "@types/uuid": "^9.0.2",
     "babel-jest": "^29.5.0",
     "eslint-config-prettier": "^8.8.0",
     "jest": "^29.5.0",
