+++ src/api/axiosCustom.ts
@@ -9,7 +9,7 @@ export const baseInstance = axios.create({
 export const authInstance = axios.create({
   baseURL: BASE_URL,
   // TODO: 토큰 관리하는 방법에 따라 변경할 수 있음
-  headers: { Authorization: `barear ${localStorage.getItem('token')}` },
+  headers: { Authorization: `Barear ${localStorage.getItem('token')}` },
 });
 
 // TODO: interceptor 재인증 로직 추가
