+++ src/pages/posts/[postId].tsx
@@ -1,6 +1,5 @@
 import { useRouter } from 'next/router';
 import React, { useState } from 'react';
-import * as S from './styles';
 import { PostType } from '@/types/post/post';
 import backIcon from '@/assets/backIcon.png';
 import noLikeIcon from '@/assets/noLike.png';
@@ -11,6 +10,7 @@ import dayjs from 'dayjs';
 import Comment from '@/components/Comment';
 import getTechImageURL from '@/utils/getTechImageUrl';
 import SendNote from '@/components/SendNote';
+import styled from '@emotion/styled';
 
 const DUMMY_DATA: PostType = {
   id: 1,
@@ -69,25 +69,25 @@ export default function Post() {
   };
 
   return (
-    <S.PostWrapper>
-      <S.BackIconWrapper onClick={() => router.push('/')}>
+    <PostWrapper>
+      <BackIconWrapper onClick={() => router.push('/')}>
         <Image src={backIcon} alt="backIcon" fill />
-      </S.BackIconWrapper>
+      </BackIconWrapper>
       <h1>{DUMMY_DATA.title}</h1>
-      <S.DescriptionWrapper>
-        <S.ProfileImage>
+      <DescriptionWrapper>
+        <ProfileImage>
           <Image src={DUMMY_DATA.user.image_url} alt="profileImage" fill />
-        </S.ProfileImage>
-        <S.OptionBox>
-          <S.OptionImage>
+        </ProfileImage>
+        <OptionBox>
+          <OptionImage>
             <Image
               src="/noteIcon.png"
               alt="noteIcon"
               fill
               onClick={() => setIsNoteOpen(true)}
             />
-          </S.OptionImage>
-          <S.OptionImage>
+          </OptionImage>
+          <OptionImage>
             {likeState ? (
               <Image
                 src={likeIcon}
@@ -103,71 +103,71 @@ export default function Post() {
                 onClick={() => setLikeState(true)}
               />
             )}
-          </S.OptionImage>
-        </S.OptionBox>
-        <S.DescriptionBox>
-          <S.Description>
+          </OptionImage>
+        </OptionBox>
+        <DescriptionBox>
+          <Description>
             <p>작성자</p>
             <text>{DUMMY_DATA.user.nickname}</text>
-          </S.Description>
-          <S.Description>
+          </Description>
+          <Description>
             <p>프로젝트명</p>
             <text>{DUMMY_DATA.project_name}</text>
-          </S.Description>
-          <S.Description>
+          </Description>
+          <Description>
             <p>생성일</p>
             <text>{dayjs(new Date()).format('YYYY.MM.DD')}</text>
-          </S.Description>
-          <S.Description>
+          </Description>
+          <Description>
             <p>모집 분야</p>
             <ul>
               {['프론트엔드', '백엔드'].map((position, idx) => (
-                <S.OptionBadge key={idx}>{position}</S.OptionBadge>
+                <OptionBadge key={idx}>{position}</OptionBadge>
               ))}
             </ul>
-          </S.Description>
-          <S.Description>
+          </Description>
+          <Description>
             <p>모집 인원</p>
             <text>{DUMMY_DATA.headcount}명</text>
-          </S.Description>
-          <S.Description>
+          </Description>
+          <Description>
             <p>모집 마감</p>
             <text>{dayjs(DUMMY_DATA.deadline).format('YYYY.MM.DD')}</text>
-          </S.Description>
-          <S.Description>
+          </Description>
+          <Description>
             <p>사용 기술</p>
             <ul>
               {DUMMY_DATA.tech_stack_list.map((tech) => (
-                <S.TechIcon key={tech.id}>
+                <TechIcon key={tech.id}>
                   <Image
                     src={getTechImageURL(tech.id) || ''}
                     alt="techIcon"
                     fill
                   />
-                </S.TechIcon>
+                </TechIcon>
               ))}
             </ul>
-          </S.Description>
-        </S.DescriptionBox>
-      </S.DescriptionWrapper>
+          </Description>
+        </DescriptionBox>
+      </DescriptionWrapper>
       <h3>
         모집마감 {dayjs(DUMMY_DATA.deadline).diff(dayjs(new Date()), 'days')}일
         남았어요!
       </h3>
-      <S.Introduce className="introduce">
+      <Introduce className="introduce">
         <h1>프로젝트 소개</h1>
         <hr />
         <div dangerouslySetInnerHTML={{ __html: DUMMY_DATA.content }} />
-      </S.Introduce>
-      <S.CommentWrapper>
+      </Introduce>
+      <CommentWrapper>
         <h1>{DUMMY_DATA.comment_list.length}개의 댓글이 있습니다.</h1>
-        <S.InputWrapper>
+        <InputWrapper>
           <textarea placeholder="내용을 입력하세요." />
-          <S.ClickButton>
+          <ClickButton>
             <Image src={enterIcon} alt="enter" fill />
-          </S.ClickButton>
-        </S.InputWrapper>
-        <S.CommentList>
+          </ClickButton>
+        </InputWrapper>
+        <CommentList>
           {DUMMY_DATA.comment_list.map((comment) => (
             <Comment
               key={comment.id}
@@ -176,9 +176,202 @@ export default function Post() {
               onClickNoteModal={onClickNoteModal}
             />
           ))}
-        </S.CommentList>
-      </S.CommentWrapper>
+        </CommentList>
+      </CommentWrapper>
       {isNoteOpen && <SendNote onClickNoteModal={onClickNoteModal} />}
-    </S.PostWrapper>
+    </PostWrapper>
   );
 }
+
+export const PostWrapper = styled.div`
+  display: flex;
+  flex-direction: column;
+  width: 100vw;
+  max-width: 1024px;
+  min-height: 100vh;
+  margin: 0 auto;
+  padding: 6rem 3rem;
+
+  & > h1 {
+    font-size: 4.275rem;
+    font-weight: bold;
+    margin: 3rem 0;
+  }
+  & > h3 {
+    align-self: center;
+    margin: 5.5rem 0;
+    font-family: var(--Noto-B);
+    font-size: 2.4rem;
+    color: ${(props) => props.theme.main_brown};
+  }
+`;
+
+export const BackIconWrapper = styled.div`
+  position: relative;
+  width: 5rem;
+  height: 5rem;
+  cursor: pointer;
+  &:hover {
+    opacity: 0.8;
+  }
+`;
+
+export const DescriptionWrapper = styled.div`
+  position: relative;
+  display: flex;
+  flex-direction: column;
+  margin-top: 4rem;
+  padding: 4.5rem;
+  background-color: #f9f8f7;
+  border-radius: 1.8rem;
+  font-size: 2.2rem;
+`;
+
+export const DescriptionBox = styled.div`
+  display: grid;
+  grid-template-columns: repeat(auto-fit, minmax(42rem, 1fr));
+  row-gap: 2.5rem;
+`;
+
+export const Description = styled.div`
+  display: flex;
+  font-family: var(--Noto-B);
+  & > p {
+    display: flex;
+    align-items: center;
+    min-width: 15rem;
+    color: ${(props) => props.theme.main_brown};
+  }
+  & > text {
+    color: #595959;
+  }
+  & > ul {
+    display: flex;
+    gap: 0.5rem;
+  }
+`;
+
+export const OptionBadge = styled.li`
+  display: flex;
+  align-items: center;
+  text-align: center;
+  border: 0.05rem solid ${(props) => props.theme.main_brown};
+  border-radius: 5px;
+  padding: 0.2rem 0.3rem;
+  color: ${(props) => props.theme.main_brown};
+  background: #f9fafb;
+  font-size: 1.5rem;
+  outline: none;
+  ${(props) => props.theme.mq.mobile} {
+    zoom: 0.5;
+  }
+`;
+
+export const ProfileImage = styled.div`
+  position: relative;
+  position: absolute;
+  top: -2.5rem;
+  left: 5rem;
+  width: 8rem;
+  height: 8rem;
+  border-radius: 50%;
+  overflow: hidden;
+`;
+export const OptionBox = styled.div`
+  display: flex;
+  align-self: flex-end;
+  gap: 2.5rem;
+  margin-bottom: 2rem;
+`;
+
+export const OptionImage = styled.div`
+  position: relative;
+  width: 2.75rem;
+  height: 2.75rem;
+  cursor: pointer;
+  &:hover {
+    opacity: 0.8;
+  }
+`;
+
+export const TechIcon = styled.div`
+  position: relative;
+  width: 3.3rem;
+  height: 3.3rem;
+  border: 1px solid ${(props) => props.theme.main_brown};
+  border-radius: 50%;
+  margin-right: 0.5rem;
+`;
+
+export const Introduce = styled.div`
+  margin-bottom: 20rem;
+  & > h1 {
+    font-family: var(--Noto-B);
+    font-size: 3.2rem;
+  }
+  & > hr {
+    width: 100%;
+    height: 3px;
+    background-color: ${(props) => props.theme.horizon_gray};
+    border: none;
+    border-radius: 3px;
+    outline: none;
+    margin: 1.5rem 0;
+  }
+  & > div {
+    h1 {
+      font-size: 2rem;
+      line-height: 2.5rem;
+    }
+  }
+`;
+
+export const CommentWrapper = styled.div`
+  & > h1 {
+    font-family: var(--Noto-B);
+    font-size: 3.2rem;
+    margin-bottom: 1.5rem;
+  }
+`;
+
+export const InputWrapper = styled.div`
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  width: 100%;
+  min-height: 7rem;
+  padding: 1rem 2rem;
+  border: 0.25rem solid ${(props) => props.theme.main_brown};
+  border-radius: 2.2rem;
+  & > textarea {
+    flex-grow: 1;
+    width: 100%;
+    outline: none;
+    border: none;
+    padding-top: 2rem;
+    font-size: 2rem;
+    resize: none;
+    background: transparent;
+    white-space: pre-wrap;
+    &:focus {
+      outline: none;
+    }
+  }
+`;
+
+export const ClickButton = styled.div`
+  position: relative;
+  width: 5rem;
+  height: 5rem;
+  cursor: pointer;
+  &:hover {
+    opacity: 0.8;
+  }
+`;
+
+export const CommentList = styled.ul`
+  display: flex;
+  flex-direction: column;
+  gap: 2.5rem;
+  margin-top: 3rem;
+`;
