+++ src/types/apiResponseType.ts
@@ -1,0 +1,4 @@
+export type ApiResponseType = {
+  code: string;
+  message: string;
+};
